fix too long ids

+++ src/main/java/ru/cultserv/adv/yandex/direct/models/campain/BannersStatItem.java
@@ -20,7 +20,7 @@ public class BannersStatItem {
      * Идентификатор объявления.
      */
     @JsonProperty("BannerID")
-    public Integer banner_id;
+    public Long banner_id;
 
     /**
      * Идентификатор фразы.
@@ -39,7 +39,7 @@ public class BannersStatItem {
      * информацию о рубрике Яндекс.Каталога.
      */
     @JsonProperty("RubricID")
-    public Integer rubric_id;
+    public Long rubric_id;
 
     /**
      * Идентификатор ретаргетинга.
@@ -47,7 +47,7 @@ public class BannersStatItem {
      * информацию о ретаргетинге.
      */
     @JsonProperty("RetargetingID")
-    public Integer retargering_id;
+    public Long retargering_id;
 
     /**
      * Ключевые слова, или название рубрики Яндекс.Каталога, или наименование условия ретаргетинга, или значение
